Restore fork upgrade test executor.
